NEW: Visualization sub-package (#9)
INFO: This package handles the creation and parsing of a JSON that can be used for the package Escher. The main module converter.py has the class JsonDictionary, which is the base for the Escher-visualization.

UPDATED: cobramod version is now 0.1.2. Most modules of the core has a proper module-documentation.This include docstings for mutiple methods and functions.. __init__.py, includes a new variable "available_databases".
FIXED: Generators in creation.py were skipping Errors. Now, they created the corresponding COBRApy objects and then yield them.
NEW: Two new errors FoundInPairError, NodeAttributeError
FIXED: Readability of extension.py upgraded. Reactions are not added to JsonDictionary if previously in model.
FIXED: Extra verbosity of parsers removed.
UPDATED: new methods for Pathway class added. This includes, add_members, __check_copy and visualize, which is the main idea of the package.
NEW: subpackage is divided in 4 modules. converter, which creates the JsonDictionary class; debug, logging configuration; items, pythonic representation of the Escher-JSON-schema; and pair, which creates the class PairDictionary, a dictionary that is linked to another one.
UPDATED: multiple test were updated and new test cases were included. The github workflow includes the new tests for the visualization.
UPDATED: depencies updated in setup.py and github workflow