Refactoring of visualization (#12)
INFO: Restructure of module graph.py and addition of vertical visualization option.
UPDATED: in __init__.py, Version pushed to 0.1.5 and function model_convert is imported from new module.
FIXED: Some Rst-domains replaced to use the corresponding one e.g from :func: to :class: in multiple modules.
UPDATED: in creation.py docstrings updated and rst-domains fixed.
REMOVED: old test with old model as it is time-expensive for the CI pipeline.
NEW: module mapping.py, This modules handles the mapping of a directed graph into a matrix that can be used in Escher. The main function is get_mapping, which checks if the graph is lineal and creates the representation in a matrix. In case of cyclic path, it will cut it.
UPDATED: whole module graph.py refactored. A new algorithm was developed to be used in the extension.py module and for the visualization sub-package. New algorithm is faster and do not create extra reactions.
UPDATED: in pathway.py, the Pathway class uses the new algorithm for its attribute "graph", which is used for its visualization. function model_convert moved to this module. Function visualize depends from the size of the mapping for the visualization size.
HOTFIX: KEGG missing in available_databases
UPDATED: Old error NotInRangeError now shows the identifier of the failed reaction in the StdOut . New exception for graph creation, GraphKeyError.
UPDATED: BiocycParser in biocyc.py got refactored in order to work with new algorithm. Method get_graph was added to BiocycParser.
UPDATED: KeggParser in kegg.py got refactored in order to work with new algorithm. Method get_graph was added to KeggParser.
REMOVED: function model_convert in utils.py moved to pathway.py.
UPDATED: in converter.py, initiation does not depeng from key-word arguments. Size of reactions is defined as 650x450 for better illustration. New attributes added for the purpose of visualize the JsonDictionary. Multiple methods includes the argumen "vertical" which indicates if the visualization should be shown vertically. Also, methods updated to adapt its behaviour depending on orientation.
NEW: extra data for test added (3 objects)
UPDATED: all tests updated to included new changes. Multiple new cases and test.
REMOVED: old sbml model removed.