FIX: non-zero-flux test uses solver tolerance (#20)
INFO: cobra_tolerance is the default value that the test will use in order to pass the tests.
HOTFIX: UserWarning should be raise if META is used in _get_gene_xml
HOTFIX: When creating genes, if name is missing, it will be changed to the identifier
UPDATED: in extension.py, multiple functions have their argument minimum removed. Docstrings modified to removed the argument 'minimum'. In test_result, demand lower_bound changed to cobra_tolerance*10. If statement changed to a try/except.
UPDATED: in test_extension.py, test updated to removed minimum. New test to check behaviour of test_result when failing
UPDATED: version bump to 0.4.1