UPDATED: Adjustment of the new github action for the merge.
UPDATED: the action is now executed on open edit reopen or close a pull request. If the pull request is closed and merged, the new version is automatically uploaded to PyPi.

UPDATED: Removed duplicate information from .toml.

UPDATED: Removed the old workflow and the file for creating the conda environment.