HOTFIX: Fix for summary order in txt format (#24)
UPDATED: extension.py now uses the suffix of the file being created to determine the format. Changes in other files are due to this adjustment.
UPDATED: summary now creates a path from string if a string was passed as filename.
UPDATED: Summary is now also tested with empty models.
UPDATED: The default behavior of the summary has been changed. it no longer creates a txt file by default, instead it does not create any file.
UPDATED: Change from the name Changes to the explicit naming of new and removed entities in the individual areas of a DataModel. This affects both the detailed and short summary. The short summary also received a new style.
UPDATED: Extension of the tests to check the separate additions and deletions.
UPDATED: version bumped to 0.5.1