New Workflow for testing, building and publishing. (#25)
NEW: Testing different python versions with current packages using tox
NEW: tox.ini this file contains the configurations for tox to perform the style checks and to run python tests using different python versions.
UPDATE: gitignore now contains all files created by tox on the fly
NEW: mypy.ini this contains the same settings that are already defined in the .toml. However, these would not be taken into account by mypy, when executed by tox.
UPDATED: Changes in summary.py and in extension.py pointed out by newer versions of black.
UPDATED: Adjustments in converter.py which are necessary due to the behavior of newer numpy versions. (An IndexError occurs with empty lists).
NEW: Github action for the automated upload of the new packet versions
UPDATED: the action is now executed on open edit reopen or close a pull request. If the pull request is closed and merged, the new version is automatically uploaded to PyPi.
UPDATED: Removed duplicate information from .toml.
UPDATED: Removed the old workflow and the file for creating the conda environment.
UPDATED: 'edited' was replaced by 'synchronize'. Edited would be triggered when the name or description of the pull request is changed. Synchronize on the other hand when there are changes in the source branch.