NEW: Database versions check
NEW: Exception for interruptions that are caused by user input and can be cached.

NEW: The BaseParser now contains functions responsible for database version control and management.

UPDATED: The parsers use the new functions to check the database versions.

UPDATED: The parsers create new folders only when the data request was successful.

UPDATED: The test set a variable of the BaseParser, so that different database versions do not interrupt the tests.