Changing the database determination
The determination of the availability of a database in BioCyc no longer takes place via the file "biocyc_db.xml" but via a cached version of this file which is downloaded directly from BioCyc.

It is now checked whether the requested database is available in the cached file or already contained in the versioning file. So online available databases as well as potentially only locally available data are retrievable.

Further adjustments that were necessary for these changes.