FIX: Bug fixes concerning the addition of cross references.
FIX: For the test cases of cross references. Here the prefix of an identifier was added and the check for add2dict_unique is no longer position dependent.

FIX: metanetx2ec can now also process List[str].

FIX: All found identifiers are now added again. There was an error that could cause some identifiers to be found but not added.