NEW: Two new entries in the documentation. One covers the credentials required for BioCyc and the other covers the new visualization options.
UPDATED: Added some DocString and used ruff.

UPDATED: The ReadtheDocs environment has changed to use mambaforge instead of the classic conda solver for the creation of the environment.

UPDATED: 'environment.yml' The Python version was set to 3.11 as autoapi is not yet compatible with 3.12 and this causes problems while creating the documentation.