Existing UI tests now also use Galata.
UPDATED: The visualization tests for Escher have been migrated to the built-in integration and now use Galata to be automatically compared to a reference. Opening created Escher maps is now no longer necessary.

UPDATED: The canvas used by Escher is now square, with the height and width corresponding to the width of the Jupyter cell.

NEW: Deprecation warning when using the original Escher integration.