Bugfixes for XRefs & Escher
BUGFIX: The XREF cache for saving locally is now always encapsulated in a list, even if only a single reference is available. This increases compatibility with the Feather file format. For loading, a new function is available that removes this encapsulation.

BUGFIX: Recent numpy changes resulted in floats being printed with an identification that they are numpy floats. This resulted in unreadable floats in the Escher integration. This has been fixed.

Bugfix: The cache for PubChem has been updated, and cids have now been correctly introduced as a list of cids instead of a separation by newlines.

BUGFIX: The Escher reference HTML for testing has been updated and works once again.

UPDATED: Test cases for the crossreferences have been updated to consider that metanetx is now disabled by default.