Adding of inchi, inchikey and smiles to Metabolites
There is now a function to add Inchi, Inchikeys and Smiles to a Cobrapy metabolite. The existing Biocyc or Chebi ID is used to query these, and to validate these entries, the Biocyc or Chebi ID is queried based on the Inchi, Inchikey or Smiles. If the IDs match and the Inchi, Inchikey or Smiles match from Biocyc and Chebi, they are added. If there is a mismatch with an existing Inchi Inchikey or Smiles, this is reported. Adding Smiles can be disabled using the settings object.